RP Charging Aura now changes color
The RP charging aura now changes colors when it's upgraded. Feel free to look over this PR if I made any mistakes ^
I'm not sure how I can fix the position and size of the aura, yet. I think it's a camera issue since the size of the aura changes depending on camera angle.